http2: expose stream oriented API
This patch adds a stream oriented API to the x/net/http2
package. It's not meant to be a final change since I have
not yet even written a proposal. I'm just posting it to
have some concrete code to be able to reference in the
discussion about the actual interface changes we want to
do.

If my posting this triggers a review request, sorry abou that.
It's my first time using gerrit.

See the discussion on github[1] for more details.

DO NOT REVIEW
DO NOT SUBMIT

[1]: https://github.com/golang/go/issues/49918#issuecomment-1138855082

Change-Id: Ifbcc6f86480a585eb92e801563708e91fe879b24